done banner table

+++ src/pages/HomePage/components/Settings/components/PostManagement/components/PostCard/components/BannerTable/index.jsx
@@ -1,11 +1,19 @@
+/* eslint-disable react/jsx-props-no-spreading */
 import { Popconfirm, Image } from 'antd';
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { connect, Link } from 'umi';
 import moment from 'moment';
+import { SortableContainer, SortableElement, SortableHandle } from 'react-sortable-hoc';
+import { MenuOutlined } from '@ant-design/icons';
 import CommonTable from '../CommonTable';
 import styles from './index.less';
 import RemoveIcon from '@/assets/homePage/removeIcon.svg';
 import EditIcon from '@/assets/homePage/editIcon.svg';
+import AddButton from '../AddButton';
+
+const DragHandle = SortableHandle(() => <MenuOutlined style={{ cursor: 'grab', color: '#999' }} />);
+const SortableItem = SortableElement((props) => <tr {...props} />);
+const SortableBody = SortableContainer((props) => <tbody {...props} />);
 
 const BannerTable = (props) => {
   const {
@@ -14,8 +22,23 @@ const BannerTable = (props) => {
     loading = false,
     refreshData = () => {},
     onEditPost = () => {},
+    onAddPost = () => {},
   } = props;
 
+  const [dataSource, setDataSource] = useState([]);
+
+  useEffect(() => {
+    const dataSourceTemp = [1, 2, 3, 4, 5].map((x) => {
+      const find = data.find((y) => y.position === x);
+      if (find) return find;
+      return {
+        position: x,
+      };
+    });
+    setDataSource(dataSourceTemp);
+  }, [JSON.stringify(data)]);
+
+  // functions
   const onDeleteAttachment = async (record) => {
     if (record?._id) {
       const res = await dispatch({
@@ -33,11 +56,17 @@ const BannerTable = (props) => {
   const getColumns = () => {
     const columns = [
       {
+        dataIndex: 'sort',
+        width: '7%',
+        align: 'center',
+        render: () => <DragHandle />,
+      },
+      {
         title: 'ID',
-        dataIndex: 'postID',
-        key: 'postID',
+        dataIndex: 'position',
+        key: 'position',
         width: '15%',
-        render: (postID) => <span className={styles.blueText}>#{postID}</span>,
+        render: (position) => <span className={styles.blueText}>{position}</span>,
       },
       {
         title: 'Media',
@@ -73,7 +102,7 @@ const BannerTable = (props) => {
         title: 'Created On',
         dataIndex: 'createdAt',
         key: 'createdAt',
-        render: (createdAt = {}) => {
+        render: (createdAt) => {
           return <span>{createdAt ? moment(createdAt).format('MM/DD/YYYY') : ''}</span>;
         },
       },
@@ -84,27 +113,76 @@ const BannerTable = (props) => {
         width: '10%',
         align: 'center',
         render: (_, record) => {
-          return (
-            <div className={styles.actions}>
-              <img src={EditIcon} alt="" onClick={() => onEditPost(record)} />
-              <Popconfirm
-                placement="left"
-                title="Are you sure?"
-                onConfirm={() => onDeleteAttachment(record)}
-              >
-                <img src={RemoveIcon} alt="" />
-              </Popconfirm>
-            </div>
-          );
+          if (record?.postID) {
+            return (
+              <div className={styles.actions}>
+                <img src={EditIcon} alt="" onClick={() => onEditPost(record)} />
+                <Popconfirm
+                  placement="left"
+                  title="Are you sure?"
+                  onConfirm={() => onDeleteAttachment(record)}
+                >
+                  <img src={RemoveIcon} alt="" />
+                </Popconfirm>
+              </div>
+            );
+          }
+          return <AddButton text="Add Banner" onClick={() => onAddPost(record)} />;
         },
       },
     ];
     return columns;
   };
 
+  const onSortEnd = async ({ oldIndex, newIndex }) => {
+    if (oldIndex !== newIndex) {
+      const find = dataSource.find((x) => x.position === oldIndex);
+      if (find) {
+        const res = await dispatch({
+          type: 'homePage/updateBannerPositionEffect',
+          payload: {
+            postID: find.postID,
+            currPosition: oldIndex,
+            newPosition: newIndex,
+          },
+        });
+        if (res.statusCode === 200) {
+          refreshData();
+        }
+      }
+    }
+  };
+
+  const DraggableContainer = (props1) => (
+    <SortableBody
+      useDragHandle
+      disableAutoscroll
+      helperClass="row-dragging"
+      onSortEnd={onSortEnd}
+      {...props1}
+    />
+  );
+
+  const DraggableBodyRow = ({ className, style, ...restProps }) => {
+    // function findIndex base on Table rowKey props and should always be a right array index
+    const index = dataSource.findIndex((x) => x.position === restProps['data-row-key']);
+    return <SortableItem index={index} {...restProps} />;
+  };
+
   return (
     <div className={styles.BannerTable}>
-      <CommonTable list={data} loading={loading} columns={getColumns()} />
+      <CommonTable
+        list={dataSource}
+        loading={loading}
+        columns={getColumns()}
+        rowKey="position"
+        components={{
+          body: {
+            wrapper: DraggableContainer,
+            row: DraggableBodyRow,
+          },
+        }}
+      />
     </div>
   );
 };
